requirements.txt
Django
PyMuPDF
python-docx
requests
    